requirements.txt
Flask
gunicorn
mysql-connector-python
requests
Pillow==10.0.1